Restructure component distribution
