+++ front-end/src/SendMessage.css
@@ -1,0 +1,4 @@
+h3 {
+  margin-top: 5px;
+  margin-bottom: 0px;
+}
