--- front-end/src/components/Message.css
@@ -1,4 +1,0 @@
-h3 {
-  margin-top: 5px;
-  margin-bottom: 0px;
-}
